+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release (dev branch)
 
 * Bug fixes:
+  * create plole-only stop, if plole has a modality not covered by matching stopos
   * Carto rendering of stop-position-only stops with bus symbol is "good" instead of "missplaced bus symbol"
   * devide Germany into smaller regions (federal states) to reduce size of Overpass queries
   * allow for larger Overpass queries